Refactor: 출력 기능을 하나의 클래스로 묶음
- PrintRanking 클래스에 있던 랭킹 출력 기능과 수익률 출력 기능을 PrintResult 클래스로 이동시킴
