+++ src/print/PrintResults.js
@@ -1,5 +1,5 @@
 const { Console } = require("@woowacourse/mission-utils");
-const { MESSAGE } = require('../utils/Constants');
+const { MESSAGE, RESULT } = require('../utils/Constants');
 
 class PrintResults {
   printLotto(amount, lottoLists) {
@@ -8,6 +8,20 @@ class PrintResults {
     lottoLists.forEach(lotto => {
       Console.print(`[${lotto.join(', ')}]`);
     });
+  };
+
+  printRanking(ranking) {
+    Console.print(RESULT.RESULT_START);
+    
+    for (let i = 0; i < 5; i++) {
+      Console.print(`${RESULT.RANKING_LIST[i]}${ranking[i]}${RESULT.RANKING_LIST_FINISH}`);
+    };
+  };
+
+  printPriz(prizePercent) {
+    const commaPrizePercent = prizePercent.replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ",");
+    Console.print(`${RESULT.BENEFIT}${commaPrizePercent}${RESULT.BENEFIT_FINISH}`);
+    Console.close();
   };
 };
