+++ src/utils/Calculate.js
@@ -1,11 +1,12 @@
 const { Console } = require("@woowacourse/mission-utils");
 const { NUMBER } = require('../utils/Constants');
 const RankingCalculate = require('./RankingCalculate');
-const PrintRanking = require('../print/PrintRanking')
+// const PrintRanking = require('../print/PrintRanking')
+const PrintResults = require('../print/PrintResults');
 
 class Calculate {
   constructor() {
-    this.printRank = new PrintRanking();
+    this.printResult = new PrintResults();
     this.ranking = [];
     this.prizeMoney = 0;
     this.earnPercentage;
@@ -14,7 +15,7 @@ class Calculate {
   rankingCalculate(lottoList, answerNumber, bonusNumber) {
     const rank = new RankingCalculate(lottoList, answerNumber, bonusNumber);
     this.ranking = rank.rankList;
-    this.printRank.showRanking(this.ranking);
+    this.printResult.printRanking(this.ranking);
   };
 
   prizeCalculate(price) {
@@ -22,7 +23,7 @@ class Calculate {
       this.prizeMoney += NUMBER.PRIZE_MONEY[i] * this.ranking[i];
     };
     this.earnPercentage = (this.prizeMoney / price * 100).toFixed(1);
-    this.printRank.showPriz(this.earnPercentage);
+    this.printResult.printPriz(this.earnPercentage);
   };
 
 };
